didn't really do much but I learned a lot which is what really matters. Should owrk perfectly for every case now, I will review backend a little more then return to front end for more review. Then I will follow up with webgpu spinning cube, and then try loading data into 3D texture!

+++ server/DTOs/VolumeDTO.cs
@@ -5,10 +5,10 @@ namespace server.DTOs;
 public class VolumeDataDto
 {
     public int[] Dimensions { get; }
-    public float[] VoxelSpacing { get; }
+    public double[] VoxelSpacing { get; }
     public byte[] VoxelData { get; }
 
-    public VolumeDataDto(int[] dimensions, float[] voxelSpacing, byte[] voxelData)
+    public VolumeDataDto(int[] dimensions, double[] voxelSpacing, byte[] voxelData)
     {
         Dimensions = dimensions;
         VoxelSpacing = voxelSpacing;
